split up base classes, de core function removes duplicate code
