+++ scvi/models/_utils.py
@@ -63,6 +63,50 @@ def scrna_raw_counts_properties(
     return properties
 
 
+def cite_seq_raw_counts_properties(
+    adata: anndata.AnnData,
+    idx1: Union[List[int], np.ndarray],
+    idx2: Union[List[int], np.ndarray],
+) -> Dict[str, np.ndarray]:
+    """
+    Computes and returns some statistics on the raw counts of two sub-populations.
+
+    Parameters
+    ----------
+    adata
+        AnnData object setup with `scvi`.
+    idx1
+        subset of indices describing the first population.
+    idx2
+        subset of indices describing the second population.
+
+    Returns
+    -------
+    type
+        Dict of ``np.ndarray`` containing, by pair (one for each sub-population),
+        mean expression per gene, proportion of non-zero expression per gene, mean of normalized expression.
+    """
+    gp = scrna_raw_counts_properties(adata, idx1, idx2)
+    protein_exp = get_from_registry(adata, _CONSTANTS.PROTEIN_EXP_KEY)
+
+    nan = np.array([np.nan] * len(adata.uns["scvi_protein_names"]))
+    protein_exp = get_from_registry(adata, _CONSTANTS.PROTEIN_EXP_KEY)
+    mean1_pro = np.asarray(protein_exp[idx1].mean(0))
+    mean2_pro = np.asarray(protein_exp[idx2].mean(0))
+    nonz1_pro = np.asarray((protein_exp[idx1] > 0).mean(0))
+    nonz2_pro = np.asarray((protein_exp[idx2] > 0).mean(0))
+    properties = dict(
+        raw_mean1=np.concatenate([gp["raw_mean1"], mean1_pro]),
+        raw_mean2=np.concatenate([gp["raw_mean2"], mean2_pro]),
+        non_zeros_proportion1=np.concatenate([gp["non_zeros_proportion1"], nonz1_pro]),
+        non_zeros_proportion2=np.concatenate([gp["non_zeros_proportion2"], nonz2_pro]),
+        raw_normalized_mean1=np.concatenate([gp["raw_normalized_mean1"], nan]),
+        raw_normalized_mean2=np.concatenate([gp["raw_normalized_mean2"], nan]),
+    )
+
+    return properties
+
+
 def _get_var_names_from_setup_anndata(adata):
     """Gets var names by checking if using raw."""
     var_names = (
